### 2017-09-12 #405 @whedon
[PRE REVIEW]: ENASearch: A Python library for interacting with ENA's API
**Submitting author:** @bebatut (<a href="http://orcid.org/0000-0001-9852-1987">Bérénice Batut</a>)
**Repository:** <a href="https://github.com/bebatut/enasearch" target ="_blank">https://github.com/bebatut/enasearch</a>
**Version:** 0.1.1
