### 2017-09-21 #405 @arfon
@brainstorm - would you be willing to edit this submission for JOSS?
