### 2017-09-25 #405 @brainstorm
@vals The reviews are really straightforward and fast. The ones I've done myself took between 10 and 20 minutes since JOSS is a highly automated github journal.

The checklist you'll be presented with when I assign you as a reviewer is very easy to go through as you review the software:

http://joss.theoj.org/about#reviewer_guidelines

No deadlines, just proceed at your own pace, but don't take too long, please. Are you in for it? ;)
