### 2017-09-25 #418 @whedon
[REVIEW]: ENASearch: A Python library for interacting with ENA's API
**Submitting author:** @bebatut (<a href="http://orcid.org/0000-0001-9852-1987">Bérénice Batut</a>)
**Repository:** <a href="https://github.com/bebatut/enasearch" target ="_blank">https://github.com/bebatut/enasearch</a>
**Version:** 0.2.0
