### 2017-10-02 bebatut/enasearch#25 @vals
No functional claims of the software
There is no description of what the software does. It does say it is an interface to the ENA API, but there are no examples or descriptions of what the use of this would be, or what can be accomplished with the software.
