### 2017-10-02 bebatut/enasearch#26 @vals
No statement of need
There are not statements of what problems the software is aiming to solve.
