### 2017-10-02 bebatut/enasearch#27 @vals
No example usage
The repository nor documentation provides an example of how the software can be used to solve a real world analysis problem.
