### 2017-10-02 bebatut/enasearch#28 @vals
The functionality documentation is not clear
While every command _is_ written out in the documentation, the functionality of the command, what the output is or does, or what the options are, are not described.

Example:
![screen shot 2017-10-02 at 11 59 19](https://user-images.githubusercontent.com/668803/31074584-3bc07e0e-a769-11e7-8712-5645907a8256.png)

Here it would be helpful to describe what an 'analysis field' is, and what the `[OPTIONS]` are. What can I use a list of analysis fields for?
