### 2017-10-02 bebatut/enasearch#29 @vals
No JOSS paper file
The repository does not contain a JOSS paper (either PDF or MD).
