### 2017-10-02 #418 @vals
I have now reviewed the repository in accordance to the guidelines.

Due to the following issues, I would not accept it at the moment:

  - No functional claims of the software: bebatut/enasearch#25
  - No statement of need: bebatut/enasearch#26
  - No example usage: bebatut/enasearch#27
  - The functionality documentation is not clear: bebatut/enasearch#28
  - No JOSS paper file: bebatut/enasearch#29

/Valentine
