### 2017-10-02 #418 @brainstorm
Thanks @vals for your thorough review 👍

Let's give a chance to @bebatut to address those issues and give some feedback ;)
