### 2017-10-03 bebatut/enasearch#29 @bebatut
Hi @vals,

The paper for JOSS is in [`src/paper/paper.md`](https://github.com/bebatut/enasearch/blob/master/src/paper/paper.md) (PR #24). Must it be directly at the root of the repo?

Bérénice
