+++ docs/_sources/index.rst.txt
@@ -1,14 +1,0 @@
-Welcome to ENASearch's documentation!
-=====================================
-
-ENASearch is a Python library for interacting with `ENA <http://www.ebi.ac.uk/ena/browse/programmatic-access>`_'s API.
-
-
-.. toctree::
-  :caption: ENASearch documentation
-  :maxdepth: 4
-
-  installation
-  cli_usage
-  api_usage
-  contributing
