+++ docs/_sources/installation.rst.txt
@@ -1,21 +1,0 @@
-Installation
-============
-
-Via pip
--------
-
-.. code-block:: bash
-
-    $ pip install enasearch
-
-Via Conda
----------
-
-.. code-block:: bash
-
-    $ conda install -c bioconda enasearch
-
-On Galaxy
----------
-
-ENASearch is also available as tool for `Galaxy <https://galaxyproject.org/>`_. It is available on the `ToolShed <https://toolshed.g2.bx.psu.edu/view/iuc/suite_enasearch/>`_.
