### 2017-10-03 13dfae0
Update tests because sequence_update error (#34)
