### 2017-10-03 ac6c801
Remove _source directory (#31)
