+++ docs/_sources/api_usage.rst.txt
@@ -1,23 +1,0 @@
-Usage within Python
-===================
-
-To use ENASearch within Python:
-
-.. code-block:: python
-
-    >>> import enasearch
-
-Functions
----------
-
-.. automodule:: enasearch
-   :members:
-   :inherited-members:
-
-Data
-----
-
-The fields and their description, the formats, etc were extracted manually and stored in `csv` file in `enasearch_data`. They were then serialized to be quickly imported in `enasearch` script.
-
-To update them, you can check the corresponding section in `Contributing`.
-
