### 2017-10-03 7b88525
Add examples of usage in the doc (#33)
