+++ src/docs/use_case.rst
@@ -1,0 +1,162 @@
+Some example of usage
+=====================
+
+Original use case
+-----------------
+
+This project started with a simple problem. We had a list of thousands of ENA's run ids with possibly interesting metagenomic datasets. We needed to select some of these datasets based on several criteria (platform used for the sequencing, library preparation strategy, number of reads and bases sequenced, etc) and download them afterwards. 
+
+The metadata are accessible via ENA. For example for the run "ERR1558694", the information can be accessed at `https://www.ebi.ac.uk/ena/data/view/ERR1558694 <https://www.ebi.ac.uk/ena/data/view/ERR1558694>`_ or programmatically via:
+
+.. code-block:: bash
+
+    $ curl "https://www.ebi.ac.uk/ena/data/warehouse/filereport?accession=ERR1558694&result=read_run"
+
+We can also choose the fields to extract
+
+.. code-block:: bash
+
+    $ curl "https://www.ebi.ac.uk/ena/data/warehouse/filereport?accession=ERR1558694&result=read_run&fields=run_accession,instrument_platform,library_strategy,read_count,fastq_ftp"
+
+    run_accession   instrument_platform library_strategy    read_count  fastq_ftp
+    ERR1558694  ILLUMINA    WGS 1259641 ftp.sra.ebi.ac.uk/vol1/fastq/ERR155/004/ERR1558694/ERR1558694_1.fastq.gz;ftp.sra.ebi.ac.uk/vol1/fastq/ERR155/004/ERR1558694/ERR1558694_2.fastq.gz
+
+We started to write these URLs with Python, but they are often difficult to build (without error, specially for the names of the fields). And the extracted information has to be integrated into downstream Python scripts. We thought that a Python library to interact with ENA's API could be useful. It could help to query ENA's database, give the list of possible fields for a type of data, download data from ENA, etc. This library could be used inside Python scripts but also through command-lines inside BASH scripts.
+
+If we come back to our initial problem, we can now use ENASearch start inside Python:
+
+.. code-block:: python
+
+    >>> import enasearch
+
+First, we would like to know the accessible metadata for a run ("read_run" type in ENA):
+
+.. code-block:: python
+
+    >>> enasearch.get_returnable_fields(result="read_run", verbose=False)
+
+    ['study_accession', 'secondary_study_accession', 'sample_accession', 'secondary_sample_accession', 'experiment_accession', 'run_accession', 'submission_accession', 'tax_id', 'scientific_name', 'instrument_platform', 'instrument_model', 'library_name', 'nominal_length', 'library_layout', 'library_strategy', 'library_source', 'library_selection', 'read_count', 'base_count', 'center_name', 'first_public', 'last_updated', 'experiment_title', 'study_title', 'study_alias', 'experiment_alias', 'run_alias', 'fastq_bytes', 'fastq_md5', 'fastq_ftp', 'fastq_aspera', 'fastq_galaxy', 'submitted_bytes', 'submitted_md5', 'submitted_ftp', 'submitted_aspera', 'submitted_galaxy', 'submitted_format', 'sra_bytes', 'sra_md5', 'sra_ftp', 'sra_aspera', 'sra_galaxy', 'cram_index_ftp', 'cram_index_aspera', 'cram_index_galaxy', 'sample_alias', 'broker_name']
+
[…]
