### 2017-10-03 bebatut/enasearch#25 @bebatut
It should be solved with #33 and #30
