### 2017-10-03 bebatut/enasearch#28 @bebatut
Thanks @vals for this feedback.

I am adding a page in the documentation about the interaction with ENA using ENASearch #35. I hope it will clarified the different type of data.

For the command you shown, there is no `[OPTIONS]` per se (except help) for this command specifically. When there is options for a command, it is described:

![screen shot 2017-10-03 at 20 29 37](https://user-images.githubusercontent.com/1842467/31142151-b675f032-a879-11e7-9f60-1ad5b118fd7e.png)

I am using a tool that automatically parse the Python file and extract the documentation (description and parameters) from the description of Click to avoid redundancy between command-line description and its documentation online. I do not know how I could remove the `[OPTIONS]` for the commands without parameters...

I will add more details about the outputs of the commands
