### 2017-10-03 fab3b20
Add documentation about interaction with ENA (#35)
