+++ src/docs/ena.rst
@@ -1,0 +1,105 @@
+Interacting with ENA Database
+=============================
+
+Data
+----
+
+Data in ENA are organzed into 11 domains (or type): 
+
++------------+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
+| Domain     | Description                                                                                                                                                                  |
++============+==============================================================================================================================================================================+
+| Assembly   | Information describing the construction of reads and sequence contigs into higher order scaffolds and chromosomes                                                            |
++------------+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
+| Sequence   | Assembled and, optionally, annotated assembled reads                                                                                                                         |
++------------+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
+| Coding     | A virtual domain comprising sequence regions reported by data providers as being protein-coding regions                                                                      |
++------------+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
+| Non-coding | A virtual domain comprising sequence regions reported by data providers as representing non-protein-coding (RNA)  genes                                                      |
++------------+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
+| Marker     | A virtual domain comprising information relating to phylogenetic, identification and molecular ecology marker  data                                                          |
++------------+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
+| Analysis   | Derived data forms, such as recalibrated aligned reads and metabarcoding identifications                                                                                     |
++------------+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
+| Read       | Raw sequencing reads from next generation platforms                                                                                                                          |
++------------+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
+| Trace      | Raw sequencing data from capillary platforms                                                                                                                                 |
++------------+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
+| Taxon      | Information relating to the organism that was the source of the sequenced biological sample                                                                                  |
++------------+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
+| Sample     | Information relating to the biological sample studied in the sequencing experiment                                                                                           |
++------------+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
+| Study      | Information relating to the scope of the sequencing effort; also known as 'Project', the primary use of study is to unite content otherwise dispersed across the ENA domains |
++------------+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
+
+Each domains are further subdivided in some cases into data classes. It is the results that can be accessed:
+
++---------------+-------------------+----------------------------------------------------------+
+| Domain        | Result            | Description                                              |
++===============+===================+==========================================================+
+| Assembly      | assembly          | Genome assemblies                                        |
++---------------+-------------------+----------------------------------------------------------+
+| Sequence      | sequence_release  | Nucleotide sequences (Release)                           |
++               +-------------------+----------------------------------------------------------+
+|               | sequence_update   | Nucleotide sequences (Update)                            |
[…]
