### 2017-10-03 #418 @bebatut
Thanks a lot @vals for the review! It helps us to improve a lot the documentation at http://bebatut.fr/enasearch/. So I hope I solved the issues you pointed:
- No functional claims of the software: it should be solved with [PR#33](https://github.com/bebatut/enasearch/pull/33) and [PR#30](https://github.com/bebatut/enasearch/pull/30)
- No statement of need: it should be solved with [PR#30](https://github.com/bebatut/enasearch/pull/30)
- No example usage: it should be solved with [PR#33](https://github.com/bebatut/enasearch/pull/33)
- The functionality documentation is not clear
   - More documentation about the interaction with ENA has been added with [PR#34](https://github.com/bebatut/enasearch/pull/35). It should clarify the different type of data and commands
   - I will add more description about the outputs of the commands during the week
- No JOSS paper file: it is located in [`src/paper/paper.md`](https://github.com/bebatut/enasearch/blob/master/src/paper/paper.md)
     Must it be at the root of the repo?

Please tell me if you think there is more missing or not clear
