### 2017-10-14 4f32f7d
Add documentation on the python commands
