+++ enasearch/__init__.py
@@ -36,10 +36,15 @@ taxonomy_results = load_object(get_data("taxonomy_results.p"))
 
 
 def get_results(verbose=True):
-    """Return the list of results in ENA as a dictionary with the key being the
-    result id and the value the result description
+    """Return the possible results (type of data) in ENA (other than taxonomy)
+
+    Each result is described in a dictionary with a description of the result, 
+    the list of returnable fields associated with the result and a dictionnary 
+    with the filter fields associated with the result
 
     :param verbose: boolean to define the printing info
+
+    :return: a dictionary with the keys being the result ids and the values dictionary to describe the results
     """
     if verbose:
         for result in results:
@@ -48,9 +53,13 @@ def get_results(verbose=True):
 
 
 def get_taxonomy_results(verbose=False):
-    """Return info about the taxonomy results
+    """Return description about the possible results accessible via the taxon portal 
+
+    Each taxonomy result is described with a short description
 
     :param verbose: boolean to define the printing info
+
+    :return: a dictionary with the keys being the result ids and the values dictionary to describe the results
     """
     if verbose:
         pprint(taxonomy_results)
@@ -58,7 +67,10 @@ def get_taxonomy_results(verbose=False):
 
 
 def check_result(result):
-    """Check if result is in the list of possible results in ENA
+    """Check if a result id is in the list of possible results accessible on ENA
+
+    This function raises an error if the result is not in the list of possible
+    results
 
     :param result: id of result to check
     """
@@ -70,7 +82,10 @@ def check_result(result):
 
 
 def check_taxonomy_result(result):
-    """Check if a result is in the list of possible taxonomz results in ENA
+    """Check if a result id is in the list of possible results in ENA Taxon Portal
+
+    This function raises an error if the result is not in the list of  possible
+    taxonomy results
 
     :param result: id of result to check
     """
@@ -82,10 +97,12 @@ def check_taxonomy_result(result):
 
 
 def get_result(result, verbose=False):
-    """Return info about a result
+    """Return the description of a result (description, returnable and filter fields)
 
     :param result: id of the result (partition of ENA db), accessible with get_results
     :param verbose: boolean to define the printing info
+
+    :return: dictionary with a description of the result, the list of returnable fields and a dictionnary with the filter fields
     """
     results = get_results(verbose=False)
     check_result(result)
@@ -96,10 +113,16 @@ def get_result(result, verbose=False):
 
 
 def get_filter_fields(result, verbose=False):
-    """Get the filter fields of a result to build a query
+    """Return the filter fields of a result
+
+    This function returns the fields that can be used to build a query on a 
+    result on ENA. Each field is described in a dictionary with a short 
+    description and its type (text, number, etc). 
 
     :param result: id of the result (partition of ENA db), accessible with get_results
     :param verbose: boolean to define the printing info
+
+    :return: dictionary with the keys being the fields ids and the values dictionary to describe the fields
     """
     result_info = get_result(result)
     filter_fields = result_info["filter_fields"]
@@ -109,10 +132,15 @@ def get_filter_fields(result, verbose=False):
 
 
 def get_returnable_fields(result, verbose=False):
-    """Get the returnable fields of a result
+    """Return the returnable fields of a result
+
+    This function returns the list of fields that can be extracted for a result
+    in a query on ENA
 
     :param result: id of the result (partition of ENA db), accessible with get_results
     :param verbose: boolean to define the printing info
+
+    :return: list of fields that can be extracted for a result
     """
     check_result(result)
     result_info = get_result(result)
@@ -123,7 +151,10 @@ def get_returnable_fields(result, verbose=False):
 
 
 def check_returnable_fields(fields, result):
-    """Check that some fields are returnable fields of a resut
[…]
