### 2017-10-16 #418 @vals
Hi @bebatut

With these changes the software is more clear. It now satisfies the conditions for a JOSS publication.

/Valentine
