### 2017-10-20 #418 @bebatut
@brainstorm I will make an archive on Zenodo. But, because of the changes I made thanks to @vals, I was thinking about increasing the version. Is it possible?
