### 2017-10-20 #418 @brainstorm
Yeah, make sure you include the changes proposed by Valentine and bump up the version ;)
