### 2017-10-20 bb2f533
Add more doc for the click commands
