### 2017-10-22 70c1878
Increase version to 0.2.0
