+++ setup.py
@@ -2,7 +2,7 @@ from setuptools import find_packages, setup
 
 setup(
     name="enasearch",
-    version="0.1.1",
+    version="0.2.0",
     author="Berenice Batut",
     author_email="berenice.batut@gmail.com",
     description=("A Python library for interacting with ENA's API"),
