+++ src/docs/conf.py
@@ -63,9 +63,9 @@ author = u'Bérénice Batut'
 # built documents.
 #
 # The short X.Y version.
-version = u'0.1.1'
+version = u'0.2.0'
 # The full version, including alpha/beta/rc tags.
-release = u'0.1.1'
+release = u'0.2.0'
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
