### 2017-10-22 afac784
Fix flake8 issue in the docstrings
