+++ enasearch/__init__.py
@@ -38,8 +38,8 @@ taxonomy_results = load_object(get_data("taxonomy_results.p"))
 def get_results(verbose=True):
     """Return the possible results (type of data) in ENA (other than taxonomy)
 
-    Each result is described in a dictionary with a description of the result, 
-    the list of returnable fields associated with the result and a dictionnary 
+    Each result is described in a dictionary with a description of the result,
+    the list of returnable fields associated with the result and a dictionnary
     with the filter fields associated with the result
 
     :param verbose: boolean to define the printing info
@@ -53,9 +53,9 @@ def get_results(verbose=True):
 
 
 def get_taxonomy_results(verbose=False):
-    """Return description about the possible results accessible via the taxon portal 
+    """Return description about the possible results accessible via the taxon portal.
 
-    Each taxonomy result is described with a short description
+    Each taxonomy result is described with a short description.
 
     :param verbose: boolean to define the printing info
 
@@ -115,9 +115,9 @@ def get_result(result, verbose=False):
 def get_filter_fields(result, verbose=False):
     """Return the filter fields of a result
 
-    This function returns the fields that can be used to build a query on a 
-    result on ENA. Each field is described in a dictionary with a short 
-    description and its type (text, number, etc). 
+    This function returns the fields that can be used to build a query on a
+    result on ENA. Each field is described in a dictionary with a short
+    description and its type (text, number, etc).
 
     :param result: id of the result (partition of ENA db), accessible with get_results
     :param verbose: boolean to define the printing info
@@ -171,8 +171,8 @@ def get_sortable_fields(result, verbose=False):
     """Return the sortable fields of a result
 
     This function returns the fields that can be used to sort the output of a
-    query for a result on ENA. Each field is described in a dictionary with a 
-    short description and its type (text, number, etc). 
+    query for a result on ENA. Each field is described in a dictionary with a
+    short description and its type (text, number, etc).
 
     :param result: id of the result (partition of ENA db), accessible with get_results
     :param verbose: boolean to define the printing info
@@ -206,7 +206,7 @@ def check_sortable_fields(fields, result):
 def get_filter_types(verbose=False):
     """Return the filters that can be used for the different type of data in a query on ENA
 
-    This function returns the filters that can be used for the different type of 
+    This function returns the filters that can be used for the different type of
     data (information available with the information on the filter fileds). For
     each type of data is given the operations applicable and a description of the
     type of expected values
@@ -352,8 +352,8 @@ def request_url(url, display, file=None):
 
     This function tooks an URL built to query or extract data from ENA and apply
     this URL. If a filepath is given, the function puts the result into the
-    file and returns the status of the request. Otherwise, the results of the 
-    request is returned by the function in different format depending of the 
+    file and returns the status of the request. Otherwise, the results of the
+    request is returned by the function in different format depending of the
     display format
 
     :param url: URL to request on ENA
@@ -387,7 +387,7 @@ def build_retrieve_url(
 ):
     """Build the URL to retrieve data or taxon
 
-    This function builds the URL to retrieve data or taxon on ENA. It takes 
+    This function builds the URL to retrieve data or taxon on ENA. It takes
     several arguments, check their validity before combining them to build the
     URL.
 
@@ -433,8 +433,7 @@ def retrieve_data(
 
     This function retrieves data (other than taxon) from ENA by:
 
-    - Building the URL based on the ids to retrieve and some parameters to 
-      format the results
+    - Building the URL based on the ids to retrieve and some parameters to format the results
     - Requesting the URL to extract the data
 
     :param ids: comma-separated identifiers for records other than Taxon
[…]
