### 2017-10-22 60f5d35
Reformate a docstring
