### 2017-10-22 0285ed1
Update the doc with functional doc from docstring
