### 2017-10-22 326ab4e
Add more functional doc (#36)
* Add documentation on the python commands

* Add more doc for the click commands

* Fix flake8 issue in the docstrings

* Reformate a docstring

* Update the doc with functional doc from docstring
