+++ enasearch/cli.py
@@ -55,81 +55,111 @@ def print_display(results, display):
 @click.group(context_settings=CONTEXT_SETTINGS)
 @click.version_option()
 def cli():
-    """ The Python library for interacting with ENA's API """
+    """The Python library for interacting with ENA's API"""
     pass
 
 
-@click.command('get_results', help="Get list of possible results")
+@click.command('get_results')
 @exception_handler
 def get_results():
-    """ Get list of possible results """
+    """Get the possible results (type of data).
+
+    This function return the possible results (or type of data) accessible with
+    ENA with their ids and a short description
+    """
     enasearch.get_results(verbose=True)
 
 
-@click.command('get_taxonomy_results', help='Get list of taxonomy results')
+@click.command('get_taxonomy_results')
 @exception_handler
 def get_taxonomy_results():
-    """ Return the list of taxonomy results in ENA """
+    """Get list of taxonomy results.
+
+    This function returns the  description about the possible results
+    accessible via the taxon portal. Each taxonomy result is described with a
+    short description"""
     taxo_results = enasearch.get_taxonomy_results(verbose=False)
     print_simple_dict(taxo_results)
 
 
-@click.command('get_filter_fields', help='Get filter fields')
+@click.command('get_filter_fields')
 @click.option(
     '--result',
     required=True,
     help='Id of a result (accessible with get_results)')
 @exception_handler
 def get_filter_fields(result):
-    """ Get the filter fields of a result to build a query """
+    """Get the filter fields of a result to build a query.
+
+    This function returns the fields that can be used to build a query on
+    a result on ENA. Each field is described on a line with field id, its
+    description, its type and to which results it is related
+    """
     fields = enasearch.get_filter_fields(result=result, verbose=False)
     print_complex_field_dict(fields)
 
 
-@click.command('get_returnable_fields', help='Get list of returnable fields')
+@click.command('get_returnable_fields')
 @click.option(
     '--result',
     required=True,
     help='Id of a result (accessible with get_results)')
 @exception_handler
 def get_returnable_fields(result):
-    """ Get the fields of a result that can returned in a report """
+    """Get the fields extractable for a result.
+
+    This function returns the fields as a list."""
     fields = enasearch.get_returnable_fields(result=result, verbose=False)
     print_list(fields)
 
 
-@click.command('get_run_fields', help='Get list of fields for a run')
+@click.command('get_run_fields')
 @exception_handler
 def get_run_fields():
-    """ Get the fields for a run """
+    """Get the fields extractable for a run.
+
+    This function returns the fields as a list."""
     fields = enasearch.get_returnable_fields(result="read_run", verbose=False)
     print_list(fields)
 
 
-@click.command('get_analysis_fields', help='Get list of fields for an analysis')
+@click.command('get_analysis_fields')
 @exception_handler
 def get_analysis_fields():
-    """ Get the fields for an analysis """
+    """Get the fields extractable for an analysis.
+
+    This function returns the fields as a list."""
     fields = enasearch.get_returnable_fields(result="analysis", verbose=True)
     print_list(fields)
 
 
-@click.command('get_sortable_fields', help='Get the sortable fields for a result')
+@click.command('get_sortable_fields')
 @click.option(
     '--result',
     required=True,
     help='Id of a result (accessible with get_results)')
 @exception_handler
 def get_sortable_fields(result):
-    """ Get the fields of a result that can sorted for a report """
+    """Get the fields of a result that can sorted.
+
+    This function returns the fields that can be used to sort the output of a
+    query for a result on ENA. Each field is described on a line with field id,
+    its description, its type and to which results it is related
+    """
     fields = enasearch.get_sortable_fields(result=result, verbose=False)
     print_complex_field_dict(fields)
 
[…]
