### 2017-10-22 7c1e280
Prepare release 0.2.0 (#38)
* Increase version to 0.2.0

* Add link to GitHub repo in doc

* Remove link to source link

* Update doc for new version
