+++ src/docs/conf.py
@@ -63,9 +63,9 @@ author = u'Bérénice Batut'
 # built documents.
 #
 # The short X.Y version.
-version = u'0.1.1'
+version = u'0.2.0'
 # The full version, including alpha/beta/rc tags.
-release = u'0.1.1'
+release = u'0.2.0'
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
@@ -104,6 +104,8 @@ html_theme_path = [sphinx_rtd_theme.get_html_theme_path()]
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
 html_static_path = ['_static']
+html_copy_source = False
+html_show_sourcelink = False
 
 # Custom sidebar templates, must be a dictionary that maps document names
 # to template names.
