### 2017-10-22 #418 @bebatut
The release is: 0.2.0
Its DOI: 10.5281/zenodo.1034619
