### 2017-10-23 #418 @brainstorm
@arfon I guess that indicating a new version does require a new DOI? I don't know what's the status of that discussion I'm afraid :-S
